Replace JSONBody reserved attr name for Body
This was done for 2 reasons: Now if you want to change the type
of the body the user can use the tag `content-type:"..."`, which
looks more like the other tags we are currently using.

The other reason is that JSONBody is unnecessarily more verbose
and in the future we would have to handle problems such as:

"what do do if the user creates 2 attributes: JSONBody and XMLBody?"

So having a simple "Body" reserved name prevents us from having
this problem.